Conflicts:
	include/stencil-composition/backend_host/backend_host.h

+++ include/common/defs.h
@@ -160,8 +160,8 @@ namespace gridtools{  namespace enumtype{
 #ifdef CXX11_ENABLED
     using int_t          = int;
     using short_t        = int;
-    using uint_t         = unsigned int;
-    using ushort_t       = unsigned short;
+    using uint_t         = int;
+    using ushort_t       = int;
     template<int_t N>
     using  static_int=boost::mpl::integral_c<int_t,N>;
     template<uint_t N>
@@ -173,8 +173,8 @@ namespace gridtools{  namespace enumtype{
 #else
     typedef int                     int_t;
     typedef int                     short_t;
-    typedef unsigned int                     uint_t;
-    typedef unsigned short                     ushort_t;
+    typedef int                     uint_t;
+    typedef int                     ushort_t;
     template<int_t N>
     struct static_int : boost::mpl::integral_c<int_t,N>{
         typedef boost::mpl::integral_c<int_t,N> type;
